backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.9.2
pymongo==4.6.3
python-dotenv==1.0.1
pillow==11.0.0
python-multipart==0.0.9
supabase==2.12.0
